sightings-map: Fix transformRequest return type
transformRequest returns a Request object, but MapLibre expects
RequestParameters. The Request type has incompatible method
property types and cannot be assigned to RequestParameters.

Return a plain object with url and signal properties instead of
a Request object. Also fix the resourceType parameter to be
optional and use maplibregl.ResourceType type.